guide, NOT cheat code
This is just an example of how the functions and loops are written

This is NOT the cheat code for the game - although I have it I don't think it is good to post it. If you really want it you can search for it but wheres the fun in that. I have to be honest, I struggled with this game. There is a one-code-fits-all spoiler out there, and even trying to break that up was hard work because I had no real documentation for the use of the robot module....in this context anyway.

I found a few very basic examples of how to arrange a basic loop, and some basic commands...and just went from there. This document will help you but it isn't a stand-alone program for all levels

The cheat code allows you to see what you need to do when writing more complex programs for your bug

The way the one-code-for-all level works (NOT THIS ONE) is like this: the bug keeps moving 90 degrees, looking in each direction, then moves when a direction is free. Sometimes it can just follow a wall around, but it is also programmed to look on every move. So if it sees a web it can turn a different direction, and if it sees a rock in the way it can pick it up, spin round, drop it, spin round and continue

That is a general idea for you to write your loops